feat(unify): rearrange settings (#20332)
* rearrange settings

* add empty states

* update tests & add new

* appease TS

* address PR comments
