+++ packages/app/src/settings/SettingsContainer.vue
@@ -5,16 +5,6 @@
   >
     <div class="space-y-24px">
       <SettingsCard
-        :title="t('settingsPage.device.title')"
-        :description="t('settingsPage.device.description')"
-        :icon="IconLaptop"
-        max-height="800px"
-      >
-        <ExternalEditorSettings :gql="props.gql" />
-        <ProxySettings :gql="props.gql" />
-        <TestingPreferences :gql="props.gql" />
-      </SettingsCard>
-      <SettingsCard
         :title="t('settingsPage.project.title')"
         :description="t('settingsPage.project.description')"
         :icon="IconFolder"
@@ -25,9 +15,27 @@
           :gql="props.gql"
         />
       </SettingsCard>
+      <SettingsCard
+        :title="t('settingsPage.device.title')"
+        :description="t('settingsPage.device.description')"
+        :icon="IconLaptop"
+        max-height="800px"
+      >
+        <ExternalEditorSettings :gql="props.gql" />
+        <ProxySettings :gql="props.gql" />
+        <TestingPreferences :gql="props.gql" />
+      </SettingsCard>
+      <SettingsCard
+        :title="t('settingsPage.cloud.title')"
+        :description="t('settingsPage.cloud.description')"
+        :icon="IconOdometer"
+        max-height="10000px"
+      >
+        <CloudSettings :gql="props.gql" />
+      </SettingsCard>
     </div>
     <hr class="border-gray-100">
-    <p class="font-light mx-auto text-center max-w-500px text-16px text-gray-500 leading-24px">
+    <p class="mx-auto font-light text-center text-gray-500 max-w-500px text-16px leading-24px">
       {{ t('settingsPage.footer.text') }}
     </p>
     <Button
@@ -49,11 +57,13 @@ import Button from '@cy/components/Button.vue'
 import ExternalEditorSettings from './device/ExternalEditorSettings.vue'
 import ProxySettings from './device/ProxySettings.vue'
 import SettingsCard from './SettingsCard.vue'
+import CloudSettings from './project/CloudSettings.vue'
 import ProjectSettings from './project/ProjectSettings.vue'
 import TestingPreferences from './device/TestingPreferences.vue'
 import type { SettingsContainerFragment } from '../generated/graphql'
 import { SettingsContainer_ReconfigureProjectDocument } from '../generated/graphql'
 import IconLaptop from '~icons/cy/laptop_x24.svg'
+import IconOdometer from '~icons/cy/object-odometer_x24.svg'
 import IconFolder from '~icons/cy/folder-outline_x24.svg'
 import SettingsIcon from '~icons/cy/settings_x16.svg'
 
@@ -69,6 +79,7 @@ gql`
 fragment SettingsContainer on Query {
   ...TestingPreferences
   ...ProjectSettings
+  ...CloudSettings
   ...ExternalEditorSettings
   ...ProxySettings
 }`
