+++ packages/app/src/settings/project/ProjectId.cy.tsx
@@ -1,3 +1,4 @@
+import { set } from 'lodash'
 import { ProjectIdFragmentDoc } from '../../generated/graphql-test'
 import ProjectId from './ProjectId.vue'
 
@@ -11,7 +12,7 @@ describe('<ProjectId />', () => {
 
     cy.mountFragment(ProjectIdFragmentDoc, {
       onResult: (result) => {
-        result.projectId = givenProjectId
+        set(result, 'currentProject.projectId', givenProjectId)
       },
       render: (gqlVal) => (
         <div class="py-4 px-8">
